resq-agent/requirements.txt
livekit-agents>=1.4.2
livekit-plugins-google>=1.4.2
python-dotenv
google-generativeai
google-cloud-aiplatform
google-cloud-storage
google-cloud-firestore
googlemaps
